Fix image shipping CICD pipeline
Fixed the docker image shipping pipeline. 

Image will now to be shipped to `ngfe/lotusim` with the tag `latest` and `<commit_ref>`

Currently kaniko is used to ship the image, in the future, crane should be used after running test on the image.

Malcom Neo <malcom.neo@fe.naval-group.com>